neuralstark/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
watchdog==4.0.1
redis==5.0.1
langchain
langchain-google-genai
pypdf
python-docx
odfdo
python-dotenv
celery==5.4.0
chromadb
langchain-community
numpy<2.0.0
langchain-chroma
sentence-transformers
langchain-huggingface
reportlab
pandas
openpyxl
xlrd
tabulate
odfdo
fastapi-cache
SpeechRecognition
pytest
